migrated apply and delete template functions to form component
