+++ src/components/dashboardComponents/ChooseDateForm.js
@@ -6,6 +6,8 @@ import {
   Heading,
   IconButton
 } from '@chakra-ui/core';
+import { useAuth } from '../../contexts/auth';
+
 
 //TODO LINK UP THE DAYS.JS TRIGGER TO MATCH THE TEMPLATE TOGGLE FROM A STATE WIDE SELECTION.
 
@@ -20,9 +22,33 @@ const ChooseDateForm = ({
   setSelected,
   templateFormOpen,
   setTemplateFormOpen,
-  applyTemplate,
-  handleDelete
+  setTemplateList,
+  templateList,
+  deleteTemplate
 }) => {
+
+  const { googleApi, api } = useAuth();
+  const handleDelete = async id => {
+    await deleteTemplate(id);
+    const templates = templateList.filter(template => template._id !== id);
+    setTemplateList(templates);
+  };
+
+  const applyTemplate = (summary, description, starttime, endtime) => {
+    //creates new date and isolates timezone offset
+    let date = new Date().toString().split("GMT");
+    //takes the first few characters of offset with + or - to be slotted in the start and end times
+    let zone = date[1].split(' ')[0].slice(0, 3);
+    const eventList = selected.map(e => ({
+      end: { dateTime: `${e}T${endtime}:00${zone}:00` },
+      start: { dateTime: `${e}T${starttime}:00${zone}:00` },
+      summary: summary,
+      description: description
+    }));
+    console.log('eventList', eventList);
+    eventList.forEach(event => api.addEvent(event));
+    setSelected([]);
+  };
 
   const [toggledTemplate, setToggledTemplate] = useState(false);
