+++ src/components/dashboardComponents/Dashboard.js
@@ -47,27 +47,7 @@ const Dashboard = ({setUserState}) => {
     })();
   }, [currentUser, formOpen]);
 
-  const handleDelete = async id => {
-    await deleteTemplate(id);
-    const templates = templateList.filter(template => template._id !== id);
-    setTemplateList(templates);
-  };
 
-  const applyTemplate = (summary, description, starttime, endtime) => {
-    //creates new date and isolates timezone offset
-    let date = new Date().toString().split("GMT");
-    //takes the first few characters of offset with + or - to be slotted in the start and end times
-    let zone = date[1].split(' ')[0].slice(0, 3);
-    const eventList = selected.map(e => ({
-      end: { dateTime: `${e}T${endtime}:00${zone}:00` },
-      start: { dateTime: `${e}T${starttime}:00${zone}:00` },
-      summary: summary,
-      description: description
-    }));
-    console.log('eventList', eventList);
-    eventList.forEach(event => api.addEvent(event));
-    setSelected([]);
-  };
 
 
   return (
@@ -95,14 +75,13 @@ const Dashboard = ({setUserState}) => {
             selected={selected}
             templateFormOpen={templateFormOpen}
             setTemplateFormOpen={setTemplateFormOpen}
-            applyTemplate={applyTemplate}
-            handleDelete={handleDelete}
 
             formOpen={formOpen}
             setFormOpen={setFormOpen}
             setTemplateList={setTemplateList}
             currentUser={currentUser}
             templateList={templateList}
+            deleteTemplate={deleteTemplate}
           />
         </Flex>
         <Box className="calendarArea" gridArea="main">
